backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.28.1
sqlite-vec==0.1.6
pydantic==2.10.4
